refactor: RecommendInput을 dto 패키지로 이동하고 빈 테스트 제거
RecommendInput은 순수 데이터 레코드라 service 패키지보다 dto 패키지가 더
어울립니다. TDD 훅이 service 패키지의 새 파일에 테스트를 강제했고 그 결과
아무것도 검증하지 않는 빈 테스트(RecommendInputTest)가 생겼습니다.

RecommendInput을 dto 패키지로 옮기고, 같은 패키지에 의존하던 ReplanAiService·
ReplanService·ReplanAiServiceTest에 명시적 import를 추가했습니다.
빈 테스트는 삭제했습니다.
